+++ reana/reana_dev/wiki.py
@@ -51,6 +51,7 @@ def create_build_status_page():
                 "reana-workflow-engine-cwl": {},
                 "reana-workflow-engine-serial": {},
                 "reana-workflow-engine-yadage": {},
+                "reana-workflow-engine-snakemake": {},
             },
         },
         "developers": {